sat-math-tutor/requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
pdf2image>=1.16.0
Pillow>=10.0.0

# OCR for English extraction
pytesseract>=0.3.10

# CLI and utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Testing
pytest>=7.0.0

